requirements.txt
# Core dependencies
fastmcp>=2.13.0,>=2.14.1,<2.15.0
pydantic>=2.0.0
pyyaml>=6.0
websockets>=12.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
structlog>=23.0.0
python-osc>=1.8.0

# Dreame vacuum integration
python-miio>=0.5.12

# Watchfiles crashproofing
watchfiles>=0.24.0

# ROS integration (optional)
rosbridge-library>=0.11.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0
matplotlib>=3.7.0

# Development dependencies (install with: pip install -e ".[dev]")
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# pytest-mock>=3.12.0
# ruff>=0.3.0
# mypy>=1.5.0
# black>=23.0.0
# isort>=5.12.0

